+++ wallets.go
@@ -21,6 +21,19 @@ type transaction struct {
 	Date              time.Time `json:"date"`
 }
 
+func postWallet(rdb *redis.Client) func(*gin.Context) {
+	return func(c *gin.Context) {
+		var receivedWallet wallet
+		if errBind := c.BindJSON(&receivedWallet); errBind != nil {
+			log.Printf("could not process received wallet, %s", errBind)
+			return
+		}
+		log.Printf("received wallet:%+v", receivedWallet)
+
+		c.IndentedJSON(http.StatusCreated, receivedWallet)
+	}
+}
+
 func postTransaction(rdb *redis.Client) func(*gin.Context) {
 	return func(c *gin.Context) {
 		var receivedTransaction transaction
@@ -28,7 +41,7 @@ func postTransaction(rdb *redis.Client) func(*gin.Context) {
 			log.Printf("could not process received transaction, %s", errBind)
 			return
 		}
-		log.Printf("received Transaction:%+v", receivedTransaction)
+		log.Printf("received transaction:%+v", receivedTransaction)
 
 		c.IndentedJSON(http.StatusCreated, receivedTransaction)
 	}